+++ src/Lorentz/Contracts/TZBTC/V1/Contract.hs
@@ -30,6 +30,7 @@ import qualified Michelson.Typed as T
 import Util.Named
 import Util.TypeTuple.Class
 
+import Lorentz.Contracts.TZBTC.Common.Doc
 import Lorentz.Contracts.TZBTC.V0 (StoreTemplateV0)
 import qualified Lorentz.Contracts.TZBTC.V0 as V0
 import qualified Lorentz.Contracts.TZBTC.V1.Impl as TZBTC
@@ -127,14 +128,7 @@ migrateStorage v1p =
 
 tzbtcDoc :: Lambda () ()
 tzbtcDoc = fakeCoercing $ do
-  -- License info
-  doc $ DComment $ mconcat
-    [ "- SP"
-    , "DX-FileCopyrightText: 2019 Bitcoin Suisse\n"
-    , "-\n"
-    , "- SP"
-    , "DX-License-Identifier: LicenseRef-MIT-BitcoinSuisse"
-    ]
+  doc licenseInfoDoc
   contractName "TZBTC" $ do
     contractGeneralDefault
     doc $ DDescription
@@ -153,12 +147,3 @@ tzbtcDoc = fakeCoercing $ do
     doc $ T.DStorageType $ DType $ Proxy @UStoreV1
     doc $ DUStoreTemplate (Proxy @(VerUStoreTemplate TZBTCv1))
     V0.tzbtcContractRaw
-
-additionalDeployNotes :: Markdown
-additionalDeployNotes =
-  "Initially originated contract has V0 which should be empty. However, it's possible\
-  \ to originate contract with some entrypoints implementation, but such origination \
-  \will highly likely exceed operation size limit, so it's recomended to originate \
-  \empty V0 contract.\n\n Once the V0 is originated, it should be upgraded to V1 in order \
-  \to be usable.\n\n The easiest way to originate and upgrade contract to V1 is to use \
-  \`tzbtc-client deployTzbtcContract` command."